requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
binance-futures-connector==1.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
fastapi==0.75.0
fastapi-utils==0.2.1
Flask==2.0.3
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomli==2.0.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
Werkzeug==2.0.3
zope.interface==5.4.0